src/pyproject.toml
[tool.poetry]
name = "hacker"
version = "1.0.0"
description = "Create a pyproject.toml file for hacker"
authors = ["ssh-22 <r.march.13_liberty0@yahoo.ne.jp>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.4"
djangorestframework = "^3.11.0"
psycopg2 = "^2.8.4"
Pillow = "^7.0.0"
django-cors-headers = "^3.2.1"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
ipython = {version = "^7.13.0"}
mypy = {version = "^0.770"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"